+++ src/opening-period/form/description/OpeningPeriodDescription.scss
