+++ src/opening-period/form/time-span/TimeInput.tsx
