+++ src/opening-period/form/time-span/TimeSpan.scss
