+++ src/opening-period/form/time-span/Weekdays.scss
